+++ GPSStyle/GPSStyle/Source/MainTabBar/TabBarModel.swift
@@ -47,16 +47,7 @@ final class TabBarModelImpl: TabBarModel {
             selectedImage: selectedImage
         )
         viewController.tabBarItem = icon
-
-        let navigationController = UINavigationController(rootViewController: viewController)
-        let appearance = UINavigationBarAppearance()
-        appearance.backgroundColor = ColorName.white.color
-        appearance.titleTextAttributes = [.foregroundColor: ColorName.black.color]
-        navigationController.navigationBar.tintColor = ColorName.black.color
-        navigationController.navigationBar.standardAppearance = appearance
-        navigationController.navigationBar.compactAppearance = appearance
-        navigationController.navigationBar.scrollEdgeAppearance = appearance
         
-        return navigationController
+        return viewController
     }
 }
