--- GPSStyle/GPSStyle/Source/Utilities/UIView+AddSubViews.swift
@@ -1,14 +1,0 @@
-//
-//  UIView+AddSubViews.swift
-//  GPSStyle
-//
-//  Created by Ivan Gorshkov on 29.10.2022.
-//
-
-import UIKit
-
-public extension UIView {
-    func addSubviews(_ subviews: UIView...) {
-        subviews.forEach(addSubview)
-    }
-}
