+++ GPSStyle/GPSStyle/Source/Utilities/UIView+Extension.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 
+// MARK: addOverlay
 extension UIView {
     func addOverlay(color: UIColor = ColorName.darkGrey.color, alpha: CGFloat = 0.1) {
         let overlay = UIView()
@@ -15,5 +16,12 @@ extension UIView {
         overlay.backgroundColor = color
         overlay.alpha = alpha
         addSubview(overlay)
+    }
+}
+
+// MARK: addSubviews
+public extension UIView {
+    func addSubviews(_ subviews: UIView...) {
+        subviews.forEach(addSubview)
     }
 }
